+++ src/components/Chat/CameraComponent.tsx
@@ -1,0 +1,79 @@
+import React, { useRef, useEffect } from 'react';
+
+interface CameraComponentProps {
+  onCapture: (dataUrl: string) => void;
+  onClose: () => void;
+}
+
+const CameraComponent: React.FC<CameraComponentProps> = ({ onCapture, onClose }) => {
+  const videoRef = useRef<HTMLVideoElement>(null);
+  const streamRef = useRef<MediaStream | null>(null);
+
+  useEffect(() => {
+    const startCamera = async () => {
+      try {
+        const stream = await navigator.mediaDevices.getUserMedia({
+          video: true,
+        });
+        streamRef.current = stream;
+
+        if (videoRef.current) {
+          videoRef.current.srcObject = stream;
+        }
+      } catch (err) {
+        console.error('Error accessing camera:', err);
+      }
+    };
+
+    startCamera();
+
+    return () => {
+      streamRef.current?.getTracks().forEach((track) => track.stop());
+      streamRef.current = null;
+    };
+  }, []);
+
+  const handleCapture = () => {
+    const video = videoRef.current;
+    if (!video) return;
+
+    const canvas = document.createElement('canvas');
+    canvas.width = video.videoWidth;
+    canvas.height = video.videoHeight;
+
+    const ctx = canvas.getContext('2d');
+    if (!ctx) return;
+
+    ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
+    onCapture(canvas.toDataURL('image/png'));
+  };
+
+  return (
+    <div className="relative w-full h-full bg-black rounded-xl overflow-hidden">
+      <video
+        ref={videoRef}
+        autoPlay
+        playsInline
+        className="absolute inset-0 w-full h-full object-cover"
+      />
+
+      <div className="absolute top-0 left-0 right-0 p-4 flex justify-end bg-gradient-to-b from-black/60 to-transparent">
+        <button
+          onClick={onClose}
+          className="w-10 h-10 rounded-full bg-black/50 text-white flex items-center justify-center hover:bg-black/70 transition"
+        >
+          ✕
+        </button>
+      </div>
+
+      <div className="absolute bottom-0 left-0 right-0 pb-6 pt-4 flex justify-center bg-gradient-to-t from-black/60 to-transparent">
+        <button
+          onClick={handleCapture}
+          className="w-16 h-16 rounded-full border-4 border-white bg-white/20 hover:bg-white/30 active:scale-95 transition"
+        />
+      </div>
+    </div>
+  );
+};
+
+export default CameraComponent;
